+++ routes/lib/socketio_handler.js
@@ -71,12 +71,20 @@ module.exports = (io, app) => {
             delete this.io._nsps[namespace];
             debug(`[${namespace}] removed`);
         },
-        addNamespace: (namespace) => {
+        setupActiveNamespace: () => {
+            const namespace = "/active";
             if (this.io._nsps[namespace] === undefined) {
                 const nsp = this.io.of(namespace);
                 nsp.on('connection', function (client) {
                     const ip = getClientIP(client);
                     debug(`[${namespace}] connection from ${ip}`);
+
+                    client.on('smartcard', (data) => {
+                        debug(`[${namespace}] smartcard ${JSON.stringify(data)} from ${ip}`);
+
+                        // TODO Emit on venue namespace?
+                        nsp.emit('smartcard', data);
+                    })
                 });
                 debug(`[${namespace}] created`);
             }
@@ -88,10 +96,6 @@ module.exports = (io, app) => {
                 nsp.on('connection', function (client) {
                     const ip = getClientIP(client);
                     debug("Client %s connected to '%s'", ip, namespace);
-
-                    client.on('get_next_match', function () {
-                        nsp.emit('venue_new_match', '');
-                    });
                 });
                 debug("Created socket.io namespace '%s'", namespace);
             }
@@ -182,10 +186,17 @@ module.exports = (io, app) => {
 
                     client.on('warmup_started', function (data) {
                         log('warmup_started');
-                        _this.io.of('/active').emit('warmup_started', { leg: data.leg, match: data.match });
-                        if (data.match.venue) {
-                            _this.io.of(`/venue/${data.match.venue.id}`).emit('warmup_started', { leg: data.leg });
-                        }
+                        axios.put(`${app.locals.kcapp.api}/leg/${data.leg.id}/warmup`)
+                            .then(() => {
+                                _this.io.of('/active').emit('warmup_started', { leg: data.leg, match: data.match });
+                                if (data.match.venue) {
+                                    _this.io.of(`/venue/${data.match.venue.id}`).emit('warmup_started', { leg: data.leg, match: data.match });
+                                }
+                            }).catch(error => {
+                                const message = `${error.message}(${error})`;
+                                debug(`[${legId}] Error when setting warmup: ${message}`);
+                                nsp.emit('error', { message: error.message, code: error.code });
+                            });
                     });
 
                     client.on('reconnect_smartboard', function (data) {
@@ -324,7 +335,7 @@ module.exports = (io, app) => {
                 });
 
                 function getNameAnnouncement(player, type) {
-                    const name = player.vocal_name === null ? player.first_name : player.vocal_name;
+                    const name = player.vocal_name === null || player.vocal_name === "" ? player.first_name : player.vocal_name;
                     if (name.includes(".wav")) {
                         const key = player.first_name.toLowerCase().replace(" ", "");
                         if (AUDIO_NAMES[key]) {