+++ src/components/spectate/spectate.component.js
@@ -1,15 +1,15 @@
 const _ = require("underscore");
-const io = require('../../../../util/socket.io-helper.js');
-const types = require('../../../../components/scorecard/components/match_types');
+const io = require('../../util/socket.io-helper.js');
+const types = require('../scorecard/components/match_types');
 
 module.exports = {
     onCreate(input) {
-        var matchName = input.match.match_mode.short_name;
+        const matchName = input.match.match_mode.short_name;
         this.state = {
             leg: input.leg,
             roundNumber: input.leg.round,
             matchName: matchName,
-            type: input.leg.leg_type.id || input.match.match_type.id,
+            matchType: input.leg.leg_type.id || input.match.match_type.id,
             submitting: false,
             socket: {},
             audioAnnouncer: undefined,
@@ -18,26 +18,39 @@ module.exports = {
     },
 
     onMount() {
-        var socket = io.connect(`${window.location.origin}/legs/${this.state.leg.id}`);
+        const socket = io.connect(`${window.location.origin}/legs/${this.state.leg.id}`);
         socket.on('score_update', this.onScoreUpdate.bind(this));
         socket.on('possible_throw', this.onPossibleThrow.bind(this));
         socket.on('leg_finished', (data) => {
-            var match = data.match;
+            const match = data.match;
+            const isVenueSpectate = this.input.venue !== undefined;
             if (match.is_finished) {
+                if (isVenueSpectate) {
+                    return;
+                }
                 location.href = `${window.location.origin}/matches/${match.id}/result`;
+            } else {
+                location.href = isVenueSpectate ? `/venues/${data.match.venue.id}/spectate` : `/matches/${match.id}/spectate`;
             }
-        });
-        socket.on('new_leg', (data) => {
-            var match = data.match;
-            location.href = `${window.location.origin}/matches/${match.id}/spectate`;
         });
         socket.on('connected', (data) => {
             socket.emit('spectator_connected', '');
         });
         socket.on('say', this.onSay.bind(this));
 
-        this.state.audioAnnouncer = new Audio();
+        if (this.input.venue) {
+            const venueSocket = io.connect(`${window.location.origin}/venue/${this.input.venue.id}`);
+            const forwardFnc = (data) => {
+                if (location.pathname.endsWith(`${data.match.id}/spectate`)) {
+                    return;
+                }
+                location.href = `/venues/${data.match.venue.id}/spectate`;
+            };
+            venueSocket.on('venue_new_match', forwardFnc);
+            venueSocket.on('warmup_started', forwardFnc);
+        }
 
+        this.state.audioAnnouncer = new Audio();
         this.state.socket = socket;
     },
 
@@ -57,8 +70,8 @@ module.exports = {
     },
 
     onPossibleThrow(data) {
-        var component = this.findActive(this.getComponents('players'));
-        if (this.state.type === types.X01) {
+        const component = this.findActive(this.getComponents('players'));
+        if (this.state.matchType === types.X01) {
             // Set current dart
             if (data.is_undo) {
                 component.getDart(data.darts_thrown).reset();
@@ -69,7 +82,7 @@ module.exports = {
             component.state.totalScore += data.score * data.multiplier;
 
             // Update player score
-            var header = this.getComponent(`player-${data.current_player_id}`);
+            const header = this.getComponent(`player-${data.current_player_id}`);
             header.state.player.current_score -= (data.score * data.multiplier)
             header.setStateDirty('player');
         } else {
@@ -86,7 +99,7 @@ module.exports = {
         if (!playerId) {
             playerId = this.findActive(this.getComponents('players')).state.playerId;
         }
-        this.getComponent('player-' + playerId).setScored(scored);
+        this.getComponent(`player-${playerId}`).setScored(scored);
     },
 
     findActive(components) {