+++ src/pages/index-controller/components/index-controller.component.js
@@ -1,6 +1,7 @@
 const _ = require('underscore');
 const axios = require('axios');
 const localStorage = require('../../../util/localstorage.js');
+const io = require('../../../util/socket.io-helper.js');
 const types = require("../../../components/scorecard/components/match_types.js")
 
 module.exports = {
@@ -16,6 +17,7 @@ module.exports = {
         this.state = {
             steps: steps,
             step: steps.INITIAL,
+            submitting: false,
             officeId: undefined,
             venueId: undefined,
             playersSelected: [],
@@ -58,14 +60,25 @@ module.exports = {
                 axios.get(`${window.location.protocol}//${window.location.hostname}${this.input.locals.kcapp.api_path}/venue/${this.state.venueId}/players`)
                     .then((ids) => {
                         const playerIds = ids.data;
-                        this.state.players = _.sortBy(_.reject(this.input.players, (player) => {
+                        this.state.players = _.reject(this.input.players, (player) => {
                             return !playerIds.includes(player.id);
-                        }), (player) => player.name);
+                        });
                         this.setStateDirty('players');
                     }).catch(error => {
                         console.log(`Error when getting recent players ${error}`);
                     });
-        }
+                const socket = io.connect(`${window.location.origin}/active`);
+                socket.on('smartcard', (data) => {
+                    if (parseInt(data.venue_id) === this.state.venueId) {
+                        const player = _.find(this.input.players, (player) => {
+                            return player.smartcard_uid === data.uid;
+                        });
+                        if (player) {
+                            this.addPlayer(null, player);
+                        }
+                    }
+                });
+            }
         }
     },
     onNewMatch(e) {
@@ -127,7 +140,7 @@ module.exports = {
         e.preventDefault();
     },
     addPlayer(event, selected) {
-        const player = selected.input.data;
+        const player = selected.input ? selected.input.data : selected;
 
         if (!this.state.playersSelected.includes(player)) {
             this.state.playersSelected.push(player);