+++ src/pages/players/components/players-table/players-table.component.js
@@ -1,6 +1,9 @@
 module.exports = {
     viewPlayer(event, element) {
-        var playerId = element.attributes['data-player'].value;
-        location.href = '/players/' + playerId + '/statistics'
+        const playerId = element.attributes['data-player'].value;
+        location.href = `/players/${playerId}/statistics`;
+    },
+    editPlayer(event, element) {
+        this.emit('edit-player', element.attributes['data-player'].value);
     }
 }